CCD-3659 refactor GlobalSearch validators: part 2
   [CCD-3659](https://tools.hmcts.net/jira/browse/CCD-3659)

Refactor GlobalSearch validators: part 2:
* update references
