+++ src/main/java/uk/gov/hmcts/ccd/domain/model/search/global/SortCriteria.java
@@ -2,8 +2,8 @@ package uk.gov.hmcts.ccd.domain.model.search.global;
 
 import lombok.Getter;
 import lombok.Setter;
-import uk.gov.hmcts.ccd.domain.model.std.validator.ValidSortBy;
-import uk.gov.hmcts.ccd.domain.model.std.validator.ValidSortDirection;
+import uk.gov.hmcts.ccd.domain.model.std.validator.globalsearch.ValidSortBy;
+import uk.gov.hmcts.ccd.domain.model.std.validator.globalsearch.ValidSortDirection;
 
 @Getter
 @Setter
