+++ src/main/java/uk/gov/hmcts/ccd/domain/model/std/validator/globalsearch/ValidSearchCriteria.java
@@ -1,4 +1,6 @@
-package uk.gov.hmcts.ccd.domain.model.std.validator;
+package uk.gov.hmcts.ccd.domain.model.std.validator.globalsearch;
+
+import uk.gov.hmcts.ccd.domain.model.std.validator.ValidationError;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
@@ -13,7 +15,6 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 @Target({PARAMETER, FIELD})
 @Retention(RUNTIME)
 public @interface ValidSearchCriteria {
-
 
     String message() default ValidationError.SEARCH_CRITERIA_MISSING;
