+++ src/main/java/uk/gov/hmcts/ccd/domain/model/std/validator/globalsearch/ValidSortDirection.java
@@ -1,4 +1,6 @@
-package uk.gov.hmcts.ccd.domain.model.std.validator;
+package uk.gov.hmcts.ccd.domain.model.std.validator.globalsearch;
+
+import uk.gov.hmcts.ccd.domain.model.std.validator.ValidationError;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
@@ -13,11 +15,11 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 @Target({PARAMETER, FIELD})
 @Retention(RUNTIME)
 public @interface ValidSortDirection {
+
     String message() default ValidationError.SORT_DIRECTION_INVALID;
 
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
-
 
 }
