+++ src/test/java/uk/gov/hmcts/ccd/domain/model/std/validator/globalsearch/SearchCriteriaValidatorTest.java
@@ -1,18 +1,22 @@
-package uk.gov.hmcts.ccd.endpoint.std.validator;
+package uk.gov.hmcts.ccd.domain.model.std.validator.globalsearch;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
 import uk.gov.hmcts.ccd.domain.model.search.global.Party;
 import uk.gov.hmcts.ccd.domain.model.search.global.SearchCriteria;
-import uk.gov.hmcts.ccd.domain.model.std.validator.SearchCriteriaValidator;
 
 import javax.validation.ConstraintValidatorContext;
 import java.util.ArrayList;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.when;
 
 class SearchCriteriaValidatorTest {
+
     @Mock
     private ConstraintValidatorContext constraintValidatorContext;
 
@@ -20,6 +24,14 @@ class SearchCriteriaValidatorTest {
     ConstraintValidatorContext.ConstraintViolationBuilder constraintViolationBuilder;
 
     private final SearchCriteriaValidator validator = new SearchCriteriaValidator();
+
+    @BeforeEach
+    void setUp() {
+        MockitoAnnotations.openMocks(this);
+
+        when(constraintValidatorContext.buildConstraintViolationWithTemplate(anyString())).thenReturn(
+            constraintViolationBuilder);
+    }
 
     @Test
     void returnsTrueWhenSearchCriteriaHasOneValidField() {
