+++ src/test/java/uk/gov/hmcts/ccd/domain/model/std/validator/globalsearch/SortByValidatorTest.java
@@ -1,14 +1,17 @@
-package uk.gov.hmcts.ccd.endpoint.std.validator;
+package uk.gov.hmcts.ccd.domain.model.std.validator.globalsearch;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
 import uk.gov.hmcts.ccd.domain.model.search.global.GlobalSearchSortByCategory;
-import uk.gov.hmcts.ccd.domain.model.std.validator.SortByValidator;
 
 import javax.validation.ConstraintValidatorContext;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.when;
 
 class SortByValidatorTest {
 
@@ -19,6 +22,14 @@ class SortByValidatorTest {
     ConstraintValidatorContext.ConstraintViolationBuilder constraintViolationBuilder;
 
     private final SortByValidator validator = new SortByValidator();
+
+    @BeforeEach
+    void setUp() {
+        MockitoAnnotations.openMocks(this);
+
+        when(constraintValidatorContext.buildConstraintViolationWithTemplate(anyString())).thenReturn(
+            constraintViolationBuilder);
+    }
 
     @Test
     void returnsTrueWhenSortByNull() {
