+++ src/test/java/uk/gov/hmcts/ccd/domain/model/std/validator/globalsearch/SortDirectionValidatorTest.java
@@ -1,14 +1,17 @@
-package uk.gov.hmcts.ccd.endpoint.std.validator;
+package uk.gov.hmcts.ccd.domain.model.std.validator.globalsearch;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
 import uk.gov.hmcts.ccd.domain.model.search.global.GlobalSearchSortDirection;
-import uk.gov.hmcts.ccd.domain.model.std.validator.SortDirectionValidator;
 
 import javax.validation.ConstraintValidatorContext;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.when;
 
 class SortDirectionValidatorTest {
 
@@ -19,6 +22,14 @@ class SortDirectionValidatorTest {
     ConstraintValidatorContext.ConstraintViolationBuilder constraintViolationBuilder;
 
     private final SortDirectionValidator validator = new SortDirectionValidator();
+
+    @BeforeEach
+    void setUp() {
+        MockitoAnnotations.openMocks(this);
+
+        when(constraintValidatorContext.buildConstraintViolationWithTemplate(anyString())).thenReturn(
+            constraintViolationBuilder);
+    }
 
     @Test
     void returnsTrueWhenSortDirectionNull() {
